SCRUM-24 FEAT: create internal payment and update its status if the payment was approved

+++ src/main/java/com/MedievalMedia/Records/PaymentRequest.java
@@ -1,5 +1,5 @@
 package com.MedievalMedia.Records;
 
-public record PaymentRequest(Double total, String currency, String method, String intent, String description, String cancelUrl, String successUrl) {
+public record PaymentRequest(BigDecimal total, String currency, String method, String intent, String description, String cancelUrl, String successUrl) {
 
 }
